Jamfile: Add python_version/py parameter
Adds a user-specifiable parameter to the Jamfile called python_version,
along with it an alias to it called py.

The Jamfile now creates subdirectories in the bin/x86_64 directory for
each python version. This required moving __init__.py out of the
bin/x86_64 folder since it should now go into the generated
subdirectories. Also, the Be shortcut was removed since there was no
longer a predefined place for it to link to.