Compile each .cpp file into its own .so
Previously, all of the .cpp files would be linked into one massive
Be.so file. Now, these .cpp files are packaged into seperate .so
files, one .so for each .cpp

This makes development faster, as you no longer will have to wait for
Be.so to compile. However, it hurts releases, since the compiled size
tripled! I am planning on adding a switch to the Jamfile for building
release builds, so that, hopefully, we can get the best of both worlds.