Fix Haiku-PyAPI crash on R1B4
Before, feedgator would fail to start on R1B4 with some error
messages.

This fixes two problems. The changes in Message.cpp address the first.
The comments there explain the changes pretty well.

The other changes make sure that lines from the bindings that specify
default parameters, such as

> .def("SetFont", &BChanelSlider::SetFont, "", py::arg("font"), py::arg("mask")=B_FONT_ALL)

work. In this case, the default argument for mask is B_FONT_ALL.
Previously on R1B4, it was getting the symbol B_FONT_ALL from the
system's version of View.h. The system's version doesn't have the
named structs, unlike the version of View.h that comes with later
versions of Haiku. This caused pybind11 to be unable to recognize that
it should use the bindings for set_font_mask defined in View.cpp.