Jamfile: Optimize build output size
The goal here is to keep builds reasonably fast for development.
Some time hopefully we'll have the ability to specify whether to build
a debug or release build. But for now, we're stuck with trying to
optimize both build time and build size simultaneously.

Coolcoder613eb proposed four build flags to add. I tested them for
their effect on build time and size.