Fix Python object inheritance problem for BLooper
Makes that a Python class subclassing BLooper is not "forgotten" as
soon as Python no longer holds any references to it, even though
C++ is still using it.

This "forgetting" is caused by Python being unaware that C++ holds a
reference to the Python subclass. So, once Python no longer holds any
references, it frees the subclass. This, in effect, removes all of the
method overrides of the base class and leads to methods of the base
class being called directly.

Part of this commit involves switching to the smart_holder branch of
pybind11. Smart holders ensure that C++ gaining a shared_pointer to
a class also increments the reference count. This turned out to be
unnecessary to fix the problem, at least in the case of BLooper.
However, it is likely that it will be necessary for some other classes
so I have decided to keep those changes.

This commit partially breaks BWindow, since it had to be switch from
py::nodelete to py::smart_holder.