Fix wrong FilterWrapper, add alias class for override
as it is impossible (or really hard) to handle hook_function in python
constructors, I propose to completely ignore this feature, so we can use
the MessageFilter class.
After some test I had to rework the new constructors because they were not
accepted with the alias class used for override. And then after testings I
realized something similar to what was adopted in BApplication. The need to
override manually because PYBIND11_OVERLOAD cannot handle the BHandler**
(pointer-to-pointer), nor it is able to handle the tuple returned by python.