Adapt constructors, fix Target function, initial docstrings
Adapt the constructors as the optional status parameter cannot be
modified in python (integers are immutable)
Thus create static definitions to create the BMessenger and return
both the messenger and the status in a tuple

Target had a wrong wrapper, now there's a mandatory python-friendly
version as it is impossible to pass a pointer to a pointer.

(more or less) half the definitions has a docstring.